Use 64-bit values for shared material value update versions.
This follows other similar systems, such as the frame counter and shader
variable group commit count. While overflow should almost certainly not be
an issue, this pretty much guarantees it will never happen.

+++ modules/Render/Render/include/DeepSea/Render/Resources/SharedMaterialValues.h
@@ -329,9 +329,9 @@ DS_RENDER_EXPORT bool dsSharedMaterialValues_clear(dsSharedMaterialValues* value
  * updates need to be processed without processing individual elements.
  *
  * @param values The shared material values to get the version from.
- * @return False if values is NULL.
+ * @return The pointer version.
  */
-DS_RENDER_EXPORT uint32_t dsSharedMaterialValues_getPointerVersion(
+DS_RENDER_EXPORT uint64_t dsSharedMaterialValues_getPointerVersion(
 	const dsSharedMaterialValues* values);
 
 /**
@@ -342,9 +342,9 @@ DS_RENDER_EXPORT uint32_t dsSharedMaterialValues_getPointerVersion(
  * individual elements.
  *
  * @param values The shared material values to get the version from.
- * @return False if values is NULL.
+ * @return The offset version.
  */
-DS_RENDER_EXPORT uint32_t dsSharedMaterialValues_getOffsetVersion(
+DS_RENDER_EXPORT uint64_t dsSharedMaterialValues_getOffsetVersion(
 	const dsSharedMaterialValues* values);
 
 /**
